### 2020-09-08 #2647 @whedon
[PRE REVIEW]: medrxivr: Accessing and searching medRxiv and bioRxiv preprint data in R
**Submitting author:** @mcguinlu (<a href="http://orcid.org/0000-0001-8730-9761">Luke A McGuinness</a>)
**Repository:** <a href="https://github.com/ropensci/medrxivr" target ="_blank">https://github.com/ropensci/medrxivr</a>
**Version:** v0.0.3
