### 2020-09-08 #2647 @mcguinlu
This package has been reviewed by rOpenSci: https://github.com/ropensci/software-review/issues/380<https://github.com/ropensci/software-review/issues/380#issue-comment-box> @ropensci/editors
