+++ NEWS.md
@@ -1,3 +1,5 @@
+# medrxivr 1.0.0
+
 # medrxivr 0.0.2
 
 Major changes: 
