+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: medrxivr
 Title: Access and Search MedRxiv and BioRxiv Preprint Data
-Version: 0.0.2.900
+Version: 1.0.0
 Authors@R: c(
     person("Luke", "McGuinness",
            role = c("aut", "cre"),
