### 2020-09-09 f593bf5
Update package version, and add details to NEWS
