+++ cran-comments.md
@@ -1,0 +1,13 @@
+## Test environments
+
+* local windows R installation, R 3.6.1
+* windows-latest (via GitHub actions), (release)
+* macOS-latest (via GitHub actions), (release)
+* ubuntu-20.04 (via GitHub actions), (release, devel)
+* win-builder (devel)
+
+## R CMD check results
+
+0 errors | 0 warnings | 1 note
+
+* This is a new release.
