+++ pkgdown.yml
@@ -2,5 +2,5 @@ pandoc: 2.7.3
 pkgdown: 1.6.0.9000
 pkgdown_sha: f3d5e4141b7893df4f0a122983f89a440d187625
 articles: []
-last_built: 2020-09-08T16:58Z
+last_built: 2020-09-09T13:37Z
 
