### 2020-09-09 a20335e
Minor updates for CRAN submission
