### 2020-09-10 1c71f39
CRAN Re-submission
