### 2020-09-14 #2651 @arfon
:wave: @mcguinlu - thanks for your submission to JOSS. Unfortunately, even though this submission came via rOpenSci, this submission is out of scope for JOSS as we classify it a 'minor' submission ([API wrapper](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)). I've asked the rOpenSci folks to update their guidelines to make it a little clearer which rOpenSci packages might be considered out of scope here.

/ cc @karthik
