### 2020-09-14 #2651 @mcguinlu
Hi @arfon,

Thanks for your comments on the package, though I'll admit I am a little surprised that the package is not considered in scope.

While the package does interact with the API, this is mainly as a precusor to searching the medRxiv database using complex search strategies and then bulk exporting the results to a .BIB file for further processing/analysis as part of a systematic literaure review. This functionality (reproducible + transparent searching, bulk export of records) is something that is not offered by either the medRxiv website or its API, or by any other package/tool, but is regularly requested by the health librarian community:

<blockquote class="twitter-tweet"><p lang="en" dir="ltr">I have pretty much given up on medRxiv, it search interface is useless &amp; unpredictable. I have many searches that I have rerun exactly the same week prior that have come back with inconsistent wacky results. No way should a search one week get 60 citations then 3000 next <a href="https://twitter.com/hashtag/medlibs?src=hash&amp;ref_src=twsrc%5Etfw">#medlibs</a></p>&mdash; Michelle Kraft (@krafty) <a href="https://twitter.com/krafty/status/1293973175244656641?ref_src=twsrc%5Etfw">August 13, 2020</a></blockquote>

<blockquote class="twitter-tweet"><p lang="en" dir="ltr">anyone have a method for bulk export from medRxiv? <a href="https://twitter.com/hashtag/systematicreviews?src=hash&amp;ref_src=twsrc%5Etfw">#systematicreviews</a> <a href="https://twitter.com/hashtag/rapidreviews?src=hash&amp;ref_src=twsrc%5Etfw">#rapidreviews</a> <a href="https://twitter.com/hashtag/ukmedlibs?src=hash&amp;ref_src=twsrc%5Etfw">#ukmedlibs</a> <a href="https://twitter.com/hashtag/medlibs?src=hash&amp;ref_src=twsrc%5Etfw">#medlibs</a> <a href="https://twitter.com/hashtag/thanks?src=hash&amp;ref_src=twsrc%5Etfw">#thanks</a></p>&mdash; (((MarshallDozier))) (@mafrado) <a href="https://twitter.com/mafrado/status/1244905414019428353?ref_src=twsrc%5Etfw">March 31, 2020</a></blockquote>

<blockquote class="twitter-tweet"><p lang="en" dir="ltr">Re-amplifying this question. I&#39;m grappling with medRxiv this morning and having issues downloading. Any ideas <a href="https://twitter.com/hashtag/medlibs?src=hash&amp;ref_src=twsrc%5Etfw">#medlibs</a> people? <a href="https://t.co/pBB8j7hy8U">https://t.co/pBB8j7hy8U</a></p>&mdash; Jane Falconer (@falkie71) <a href="https://twitter.com/falkie71/status/1253614501733249024?ref_src=twsrc%5Etfw">April 24, 2020</a></blockquote>

In addition, the main functionality of the package was developed before the medRxiv API came online - previously `medrxivr` used webscraping to import the data (for example, see this [commit point from back in February](https://github.com/ropensci/medrxivr/blob/71046e6a2d847cb529d351295cedfacba18d6b6b)), which is why I think the designation as an "API wrapper" might not be accurate.

Finally, just to make you aware that the package has already been used in two published academic publications that I know of (see [here](https://www.thelancet.com/journals/laninf/article/PIIS1473-3099(20)30630-7/fulltext) and [here](https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0234765)) (COI: I am one of the co-authors on the second paper)

I fully respect your decision as EiC, but just wanted to make sure that the decision wasn't due to an error on my part, in that I may not have accurately described `medrxivr` in the paper I submitted.

Thanks for your continued consideration!
