### 2020-09-17 f6dd77f
Wrap package names, software names and API (application programming interface) names in single quotes in title and description
