+++ DESCRIPTION
@@ -19,10 +19,10 @@ Description: An increasingly important source of health-related bibliographic
     health-related sciences are medRxiv <https://www.medrxiv.org/> and
     bioRxiv <https://www.biorxiv.org/>, both of which are operated by the Cold 
     Spring Harbor Laboratory. medrxivr provides programmatic access to the 
-    Cold Spring Harbour Laboratory (CSHL) API <https://api.biorxiv.org/>,
+    'Cold Spring Harbour Laboratory (CSHL) API' <https://api.biorxiv.org/>,
     allowing users to easily download medRxiv and bioRxiv preprint metadata
     (e.g. title, abstract, publication date, author list, etc) into R. 
-    medrxivr also provides functions to search the downloaded preprint records 
+    'medrxivr' also provides functions to search the downloaded preprint records 
     using regular expressions and Boolean logic, as well as helper functions 
     that allow users to export their search results to a .BIB file for easy 
     import to a reference manager and to download the full-text PDFs of 
