### 2020-09-17 ae343f2
Remove examples for unexported functions
