+++ man/clean_api_df.Rd
@@ -15,9 +15,4 @@ Cleaned dataframe
 \description{
 Helper script to clean data from API to make it compatible with mx_search()
 }
-\examples{
-\dontrun{
-df <- clean_api_df(df)
-}
-}
 \keyword{internal}
