### 2020-09-17 168c9d2
Remove CRAN checks badge
