### 2020-09-17 3c520ec
Update to address new CRAN feedback
