+++ pkgdown.yml
@@ -1,6 +1,9 @@
 pandoc: 2.7.3
-pkgdown: 1.6.0.9000
-pkgdown_sha: f3d5e4141b7893df4f0a122983f89a440d187625
-articles: []
-last_built: 2020-09-10T11:05Z
+pkgdown: 1.6.1.9000
+pkgdown_sha: cc9c7b1be8573f8af3453f6e5d0422d86c916701
+articles:
+  building-complex-search-strategies: building-complex-search-strategies.html
+  medrxiv-api: medrxiv-api.html
+  medrxivr: medrxivr.html
+last_built: 2020-09-17T10:09Z
 
