### 2020-09-21 #2651 @mcguinlu
Good morning @arfon (& @danielskatz, for reference ), and happy Monday!

As it's been a week, just flagging the above again in case it got lost in the mix - would be great to hear what you think, particularly re: comparison with `chirps`.
