+++ man/mx_api_doi.Rd
@@ -29,7 +29,7 @@ Provides programmatic access to data on a single preprint
   identified by a unique Digital Object Identifier (DOI).
 }
 \examples{
-\dontrun{
+\donttest{
 mx_data <- mx_api_doi("10.1101/2020.02.25.20021568")
 }
 }
