### 2020-09-24 2de04e8
Change example vignettes and tests for mx_download() and mx_export() to use tempfile and tempdir rather than modifying a users home filespace.
