+++ tests/testthat/test-download.R
@@ -10,37 +10,41 @@ mx_result <-
     doi = "10.1101/19003301"
   )
 
+tmpdir <- tempdir()
+
 test_that("Inital output", {
   skip_on_cran()
   skip_if_offline()
-  expect_message(mx_download(mx_result, "pdf"), regexp = "Downloading")
+  expect_message(mx_download(mx_result, tmpdir), regexp = "Downloading")
 })
 
 test_that("Already downloaded", {
   skip_on_cran()
   skip_if_offline()
-  expect_message(mx_download(mx_result, "pdf"), regexp = "downloaded")
+  expect_message(mx_download(mx_result, tmpdir), regexp = "downloaded")
 })
 
 test_that("Naming of downloaded PDFs", {
   skip_on_cran()
   skip_if_offline()
-  mx_download(mx_result, "pdf", name = "ID")
-  expect_equal(file.exists("pdf/271.pdf"), TRUE)
-  mx_download(mx_result, "pdf", name = "DOI")
-  expect_equal(file.exists("pdf/10.1101_19003301.pdf"), TRUE)
+  mx_download(mx_result, tmpdir, name = "ID")
+  expect_equal(file.exists(paste0(tmpdir,"/271.pdf")), TRUE)
+  mx_download(mx_result, tmpdir, name = "DOI")
+  expect_equal(file.exists(paste0(tmpdir,"/271_10.1101_19003301.pdf")), TRUE)
 })
 
-if (dir.exists("pdf") == TRUE) {
-  unlink("pdf", recursive = TRUE)
-}
+mx_result <-
+  data.frame(
+    link_pdf = paste0("https://www.medrxiv.org/content/",
+                      "10.1101/2020.09.23.20197558v1.full.pdf"),
+             ID = "272",
+             doi = "10.1101/2020.09.23.20197558")
 
 test_that("Status update", {
   skip_on_cran()
   skip_if_offline()
-  expect_message(mx_download(mx_result, "pdf", print_update = 1), regexp = "%")
+  expect_message(mx_download(mx_result,
+                             tmpdir,
+                             print_update = 1),
+                 regexp = "%")
 })
-
-if (dir.exists("pdf") == TRUE) {
-  unlink("pdf", recursive = TRUE)
-}
