### 2020-09-24 4f2835c
Update CRAN comments to describe changes
