+++ cran-comments.md
@@ -2,9 +2,16 @@
 This is a resubmission. In this version I have:
 
 * Wrapped package names, software names and API names in single quotes in title
-and description.
+and description (Sorry that I missed one last time!)
 
-* Removed the examples for unexported functions, as requested.
+* Replaced \dontrun with \donttest in all examples, as requested.
+
+* Replaced the long-running example so that only a single PDF is downloaded.
+
+* Ensured that examples/vignettes do not modify the user's home filespace in
+the examples, vignettes, and tests. All previous hardcoded file and directory
+locations have been replaced using tempfile() and tempdir(), respectively.
+
 
 ## Test environments
 
