+++ R/mx_api.R
@@ -25,7 +25,8 @@
 #'
 #' @examples
 #' \donttest{
-#' mx_data <- mx_api_content()
+#' mx_data <- mx_api_content(from_date = "2020-01-01",
+#' to_date = "2020-01-07")
 #' }
 #' @importFrom dplyr %>%
 #' @importFrom rlang .data
