+++ man/mx_api_content.Rd
@@ -42,7 +42,8 @@ Provides programmatic access to all preprints available through
 }
 \examples{
 \donttest{
-mx_data <- mx_api_content()
+mx_data <- mx_api_content(from_date = "2020-01-01",
+to_date = "2020-01-07")
 }
 }
 \seealso{
