### 2020-09-24 3b7ede7
Update NEWS.md
