+++ NEWS.md
@@ -1,3 +1,5 @@
+# medrxivr (development version)
+
 # medrxivr 0.0.3
 
 Version created for submission to JOSS and CRAN, and onboarded to rOpenSci following peer-review. 
