+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: medrxivr
 Title: Access and Search MedRxiv and BioRxiv Preprint Data
-Version: 0.0.3
+Version: 0.0.3.9000
 Authors@R: c(
     person("Luke", "McGuinness",
            role = c("aut", "cre"),
