### 2020-09-30 a8f85ae
Add CRAN badges and update install instructions
