### 2020-10-07 afa5d27
Delete pkgdown.yml
Fixes #11 (I hope)
