### 2020-10-07 #2651 @danielskatz
@mcguinlu asked me to reopen this so that he could further discuss it with @arfon and the JOSS team, so I am doing so
