### 2020-10-07 #2651 @mcguinlu
Thanks @danielskatz!

Also just flagging this to @maelle, as I know she has done a lot of work recently on updating the rOpenSci dev guide, and I agree with @arfon's point above that it might be useful to have clear guidelines on the types of rOpenSci packages that are considered out of scope by JOSS included in the rOpenSci dev guide.
