### 2020-10-07 #2651 @karthik
Thanks @mcguinlu. We have started a discussion to update the dev guide so it clearly reflects what might be out of scope for JOSS.
