### 2020-10-08 #2651 @danielskatz
This package has been reviewed by rOpenSci: https://github.com/ropensci/software-review/issues/380
