### 2020-10-08 #2651 @danielskatz
@mcguinlu - can you merge https://github.com/ropensci/medrxivr/pull/12 or let me know what part of it you disagree with?
