### 2020-10-08 #2651 @danielskatz
Also, is this code in zenodo or another archive?  If so, what is the DOI for it there?  In https://github.com/ropensci/software-review/issues/380#issuecomment-689011508 it looks like that item is checked off...
