### 2020-10-09 #2651 @mcguinlu
Hi @danielskatz,

I've merged the changes you made in ropensci/medrxivr#12 into the main branch now. And the code for this package is archived by Zenodo. The Zenodo DOI is [10.5281/zenodo.3860023](https://doi.org/10.5281/zenodo.3860023)
