### 2020-10-09 #2651 @danielskatz
In the zenodo archive, can you change the metadata so that the title and authors match those of the JOSS paper?
