### 2020-10-09 #2651 @mcguinlu
Done - title and authors are now consistent between the JOSS paper and the Zenodo repo.
